separate connection from main code
